De-duplicate shared fixtures in the provider tests
