+++ tests/providers/fastmcp_server/test_add_to_queue.py
@@ -22,14 +22,6 @@ from music_assistant.providers.fastmcp_server.tools.queue import build_queue_ser
 
 
 @pytest.fixture
-def mounted_queue(mock_mass: Any) -> FastMCP:
-    """Build a root FastMCP with the queue sub-server mounted."""
-    mcp: FastMCP = FastMCP(name="test")
-    mcp.mount(build_queue_server(mock_mass), namespace="queue")
-    return mcp
-
-
-@pytest.fixture
 def mounted_queue_no_delete(mock_mass: Any) -> FastMCP:
     """Queue server with delete:queue permission disabled."""
     mcp: FastMCP = FastMCP(name="test")
