+++ tests/providers/fastmcp_server/test_library_list_tools.py
@@ -8,16 +8,6 @@ from typing import Any
 import pytest
 from fastmcp import Client, FastMCP
 
-from music_assistant.providers.fastmcp_server.tools.library import build_library_server
-
-
-@pytest.fixture
-def library_server(mock_mass: Any) -> FastMCP:
-    """Mount only the library sub-server."""
-    mcp: FastMCP = FastMCP(name="t")
-    mcp.mount(build_library_server(mock_mass), namespace="library")
-    return mcp
-
 
 class TestListToolsRequestFullItems:
     """
